Extract a couple more components.
