+++ src/components/index.js
@@ -3,10 +3,14 @@ import Map from './map'
 import Footer from './footer'
 import Nav from './nav'
 import Session from './session'
+import Layout from './layout'
+import Event from './event'
 
 export {
   Map,
   Footer,
   Nav,
-  Session
+  Session,
+  Layout,
+  Event
 }
